feat(web): merge topology into devices as a view toggle (IA restructure ③)
/topology was a separate top-level nav item, but 2 of its 3 views (list-by-
type, subnet-grouping) were just alternate groupings of the /devices table —
only the radial L2 graph (which needs /topology API edges) was unique.

- New DeviceTopologyView.svelte component encapsulates the radial tree
  (fetches /topology?network_id=, builds the ECharts tree, theme-aware via
  CSS-var resolution). Only the graph view was carried over; the redundant
  list/subnet views were dropped (the device table already covers them).
- /devices gains a List/Topology segmented-control toggle in the header
  (pill style, distinct from the underline sub-tab bar above it). Topology
  view shares the network filter and hides the table-only chrome (filter
  bar / pagination / bulk / column-picker). URL-synced via ?view=topology.
- /topology route now redirects to /devices?view=topology (old links/book-
  marks survive).
