+++ web/messages/en.json
@@ -145,6 +145,8 @@
 		"Tab Scan Tasks": "Scan Tasks",
 		"Tab Scan Results": "Scan Results",
 		"Tab Passive Discovery": "Passive Discovery",
+		"View List": "List",
+		"View Topology": "Topology",
 		"Camera": "Camera",
 		"Search placeholder": "Search name / IP / MAC…",
 		"Alive": "Alive",
